Extract CoverageProtein from Protein, introduce ProteinViewBean
